Move tests into apps and fix Travis config
