+++ .travis.yml
@@ -9,7 +9,7 @@ dist: xenial
 # setup environment
 env:
   global:
-    - DJANGO_SETTINGS_MODULE='MacPorts.settings'
+    - DJANGO_SETTINGS_MODULE='settings'
     - DB_HOST=localhost
     - DB_PASSWORD=
     - DB_NAME=mp
