+++ app/categories/templates/categories/exceptions/category_not_found.html
@@ -1,4 +1,4 @@
-{% extends 'ports/../../../../templates/layout.html' %}
+{% extends 'layout.html' %}
 
 {% block title %}Category Not Found |{% endblock %}
