+++ app/port_detail/templates/port_detail/exceptions/port_not_found.html
@@ -1,4 +1,4 @@
-{% extends 'ports/../../../../templates/layout.html' %}
+{% extends 'layout.html' %}
 
 {% block title %}Port Not Found |{% endblock %}
