+++ app/templates/400.html
@@ -1,4 +1,4 @@
-{% extends 'ports/layout.html' %}
+{% extends 'layout.html' %}
 
 {% block title %}400 Bad Request |{% endblock %}
