+++ app/templates/404.html
@@ -1,4 +1,4 @@
-{% extends 'ports/layout.html' %}
+{% extends 'layout.html' %}
 
 {% block title %}404 Not Found |{% endblock %}
