+++ app/templates/500.html
@@ -1,4 +1,4 @@
-{% extends 'ports/layout.html' %}
+{% extends 'layout.html' %}
 
 {% block title %}500 Internal Server Error |{% endblock %}
