+++ app/templates/about.html
@@ -1,4 +1,4 @@
-{% extends 'ports/layout.html' %}
+{% extends 'layout.html' %}
 
 {% block title %}About | {% endblock %}
 {% block content %}
